+++ src/transformers/models/gemma3n/modeling_gemma3n.py
@@ -1411,6 +1411,12 @@ class Gemma3nPreTrainedModel(PreTrainedModel):
         if hasattr(module, "gradient_clipping"):
             init.constant_(module.gradient_clipping, self.config.gradient_clipping)
 
+    def get_per_layer_input_embeddings(self):
+        return self.base_model.embed_tokens_per_layer
+
+    def set_per_layer_input_embeddings(self, value):
+        self.base_model.embed_tokens_per_layer = value
+
     def resize_token_embeddings(
         self,
         new_num_tokens: int | None = None,
@@ -1431,16 +1437,17 @@ class Gemma3nPreTrainedModel(PreTrainedModel):
         pad_to_multiple_of: int | None = None,
         mean_resizing: bool = True,
     ):
-        self.config.vocab_size_per_layer_input = self.vocab_size
-        embed_tokens_per_layer = self.base_model.embed_tokens_per_layer
-        new_embeddings_per_layer = self._get_resized_embeddings(
-            embed_tokens_per_layer, new_num_tokens, pad_to_multiple_of, mean_resizing
-        )
-        if hasattr(embed_tokens_per_layer, "_hf_hook"):
-            hook = embed_tokens_per_layer._hf_hook
-            add_hook_to_module(new_embeddings_per_layer, hook)
-        new_embeddings_per_layer.requires_grad_(embed_tokens_per_layer.weight.requires_grad)
-        self.base_model.embed_tokens_per_layer = new_embeddings_per_layer
+        self.config.get_text_config().vocab_size_per_layer_input = self.vocab_size
+        if self.config.get_text_config().hidden_size_per_layer_input:
+            embed_tokens_per_layer = self.get_per_layer_input_embeddings()
+            new_embeddings_per_layer = self._get_resized_embeddings(
+                embed_tokens_per_layer, new_num_tokens, pad_to_multiple_of, mean_resizing
+            )
+            if hasattr(embed_tokens_per_layer, "_hf_hook"):
+                hook = embed_tokens_per_layer._hf_hook
+                add_hook_to_module(new_embeddings_per_layer, hook)
+            new_embeddings_per_layer.requires_grad_(embed_tokens_per_layer.weight.requires_grad)
+            self.set_per_layer_input_embeddings(new_embeddings_per_layer)
 
 
 class Gemma3nAudioEncoder(Gemma3nPreTrainedModel):
@@ -2164,6 +2171,12 @@ class Gemma3nModel(Gemma3nPreTrainedModel):
             audio_hidden_states=audio_features if input_features is not None else None,
         )
 
+    def get_per_layer_input_embeddings(self):
+        return self.language_model.embed_tokens_per_layer
+
+    def set_per_layer_input_embeddings(self, value):
+        self.language_model.embed_tokens_per_layer = value
+
     @can_return_tuple
     @auto_docstring(custom_intro="Projects the last hidden state from the audio encoder into language model space.")
     def get_audio_features(
@@ -2185,20 +2198,6 @@ class Gemma3nModel(Gemma3nPreTrainedModel):
         audio_outputs.pooler_output = audio_embeds
 
         return audio_outputs
-
-    def resize_token_embeddings(
-        self,
-        new_num_tokens: int | None = None,
-        pad_to_multiple_of: int | None = None,
-        mean_resizing: bool = True,
-    ) -> nn.Embedding:
-        inputs_embeds = super().resize_token_embeddings(
-            new_num_tokens=new_num_tokens,
-            pad_to_multiple_of=pad_to_multiple_of,
-            mean_resizing=mean_resizing,
-        )
-        # TODO: fix resizing for embeds per-layer by filtering out mm-soft-tokens
-        return inputs_embeds
 
 
 @auto_docstring(
@@ -2384,19 +2383,11 @@ class Gemma3nForConditionalGeneration(Gemma3nPreTrainedModel, GenerationMixin):
 
         return model_inputs
 
-    def resize_token_embeddings(
-        self,
-        new_num_tokens: int | None = None,
-        pad_to_multiple_of: int | None = None,
-        mean_resizing: bool = True,
-    ) -> nn.Embedding:
-        inputs_embeds = super().resize_token_embeddings(
-            new_num_tokens=new_num_tokens,
-            pad_to_multiple_of=pad_to_multiple_of,
-            mean_resizing=mean_resizing,
-        )
-        # TODO: fix resizing for embeds per-layer by filtering out mm-soft-tokens
-        return inputs_embeds
+    def get_per_layer_input_embeddings(self):
+        return self.model.get_per_layer_input_embeddings()
+
+    def set_per_layer_input_embeddings(self, value):
+        self.model.set_per_layer_input_embeddings(value)
 
 
 __all__ = [
