+++ src/transformers/models/gemma3n/modular_gemma3n.py
@@ -1663,6 +1663,12 @@ class Gemma3nPreTrainedModel(Gemma2PreTrainedModel):
         if hasattr(module, "gradient_clipping"):
             init.constant_(module.gradient_clipping, self.config.gradient_clipping)
 
+    def get_per_layer_input_embeddings(self):
+        return self.base_model.embed_tokens_per_layer
+
+    def set_per_layer_input_embeddings(self, value):
+        self.base_model.embed_tokens_per_layer = value
+
     def resize_token_embeddings(
         self,
         new_num_tokens: int | None = None,
@@ -1683,16 +1689,17 @@ class Gemma3nPreTrainedModel(Gemma2PreTrainedModel):
         pad_to_multiple_of: int | None = None,
         mean_resizing: bool = True,
     ):
-        self.config.vocab_size_per_layer_input = self.vocab_size
-        embed_tokens_per_layer = self.base_model.embed_tokens_per_layer
-        new_embeddings_per_layer = self._get_resized_embeddings(
-            embed_tokens_per_layer, new_num_tokens, pad_to_multiple_of, mean_resizing
-        )
-        if hasattr(embed_tokens_per_layer, "_hf_hook"):
-            hook = embed_tokens_per_layer._hf_hook
-            add_hook_to_module(new_embeddings_per_layer, hook)
-        new_embeddings_per_layer.requires_grad_(embed_tokens_per_layer.weight.requires_grad)
-        self.base_model.embed_tokens_per_layer = new_embeddings_per_layer
+        self.config.get_text_config().vocab_size_per_layer_input = self.vocab_size
+        if self.config.get_text_config().hidden_size_per_layer_input:
+            embed_tokens_per_layer = self.get_per_layer_input_embeddings()
+            new_embeddings_per_layer = self._get_resized_embeddings(
+                embed_tokens_per_layer, new_num_tokens, pad_to_multiple_of, mean_resizing
+            )
+            if hasattr(embed_tokens_per_layer, "_hf_hook"):
+                hook = embed_tokens_per_layer._hf_hook
+                add_hook_to_module(new_embeddings_per_layer, hook)
+            new_embeddings_per_layer.requires_grad_(embed_tokens_per_layer.weight.requires_grad)
+            self.set_per_layer_input_embeddings(new_embeddings_per_layer)
 
 
 class Gemma3nAudioEncoder(Gemma3nPreTrainedModel):
@@ -2036,6 +2043,12 @@ class Gemma3nModel(PaliGemmaModel):
         self.embed_vision = Gemma3nMultimodalEmbedder(config.vision_config, config.text_config)
         self.embed_audio = Gemma3nMultimodalEmbedder(config.audio_config, config.text_config)
 
+    def get_per_layer_input_embeddings(self):
+        return self.language_model.embed_tokens_per_layer
+
+    def set_per_layer_input_embeddings(self, value):
+        self.language_model.embed_tokens_per_layer = value
+
     @can_return_tuple
     @auto_docstring(custom_intro="Projects the last hidden state from the vision model into language model space.")
     def get_image_features(
@@ -2261,20 +2274,6 @@ class Gemma3nModel(PaliGemmaModel):
 
         return audio_outputs
 
-    def resize_token_embeddings(
-        self,
-        new_num_tokens: int | None = None,
-        pad_to_multiple_of: int | None = None,
-        mean_resizing: bool = True,
-    ) -> nn.Embedding:
-        inputs_embeds = super().resize_token_embeddings(
-            new_num_tokens=new_num_tokens,
-            pad_to_multiple_of=pad_to_multiple_of,
-            mean_resizing=mean_resizing,
-        )
-        # TODO: fix resizing for embeds per-layer by filtering out mm-soft-tokens
-        return inputs_embeds
-
 
 @auto_docstring(
     custom_intro="""
@@ -2283,6 +2282,12 @@ class Gemma3nModel(PaliGemmaModel):
     """
 )
 class Gemma3nForConditionalGeneration(PaliGemmaForConditionalGeneration):
+    def get_per_layer_input_embeddings(self):
+        return self.model.get_per_layer_input_embeddings()
+
+    def set_per_layer_input_embeddings(self, value):
+        self.model.set_per_layer_input_embeddings(value)
+
     @can_return_tuple
     @auto_docstring
     def forward(
@@ -2449,20 +2454,6 @@ class Gemma3nForConditionalGeneration(PaliGemmaForConditionalGeneration):
 
     def create_masks_for_generate(self, **super_kwargs):
         raise AttributeError("Do not inherit create_masks_for_generate from PaliGemma")
-
-    def resize_token_embeddings(
-        self,
-        new_num_tokens: int | None = None,
-        pad_to_multiple_of: int | None = None,
-        mean_resizing: bool = True,
-    ) -> nn.Embedding:
-        inputs_embeds = super().resize_token_embeddings(
-            new_num_tokens=new_num_tokens,
-            pad_to_multiple_of=pad_to_multiple_of,
-            mean_resizing=mean_resizing,
-        )
-        # TODO: fix resizing for embeds per-layer by filtering out mm-soft-tokens
-        return inputs_embeds
 
 
 __all__ = [
